fdisk: fix printf stuff
Unfortunately, fdisk_warn/info/.. function was not marked by
printf __attribute__.

We don't want to break gettext stuff now, so all compiler warnings
have been fixed by casts. This is temporary solution, after release it
will be necessary to fix all the strings.

+++ fdisks/fdisk.c
@@ -200,7 +200,9 @@ void list_disk_geometry(struct fdisk_context *cxt)
 					   | SIZE_SUFFIX_3LETTER, bytes);
 
 	fdisk_colon(cxt,	_("Disk %s: %s, %llu bytes, %llu sectors"),
-			cxt->dev_path, strsz, bytes, cxt->total_sectors);
+			cxt->dev_path, strsz,
+			(unsigned long long) bytes,
+			(unsigned long long) cxt->total_sectors);
 	free(strsz);
 
 	if (fdisk_require_geometry(cxt) || fdisk_context_use_cylinders(cxt))
